feat: add ingredient model
