+++ lib/data/model/ingredient.dart
@@ -1,0 +1,11 @@
+class Ingredient {
+  final String title;
+  final String imageUrl;
+  final int quantity;
+
+  Ingredient({
+    required this.title,
+    required this.imageUrl,
+    required this.quantity,
+  });
+}
